niter: Use ranger to query ctz range.
When niter runs after the copy-header pass it sometimes fails to
simplify assumptions in a ctz loop.
As the assumption is a simple nonzero test here we can have
ranger get us the range of the shifted expression, then verify that
this range is nonzero.

This helps recognize a ctz loop in 502.gcc's compute_transp.

	PR/tree-optimization 122207

gcc/ChangeLog:

	* tree-ssa-loop-niter.cc (shifted_range_nonzero_p): New
	function.
	(number_of_iterations_cltz): Call new function.
	* tree-ssa-loop.cc (pass_scev_cprop::execute): Enable ranger.

gcc/testsuite/ChangeLog:

	* gcc.dg/tree-ssa/ctz-char.c: Remove -fno-tree-ch.
	* gcc.dg/tree-ssa/ctz-complement-char.c: Ditto.
	* gcc.dg/tree-ssa/ctz-complement-int.c: Ditto.
	* gcc.dg/tree-ssa/ctz-complement-long-long.c: Ditto.
	* gcc.dg/tree-ssa/ctz-complement-long.c: Ditto.
	* gcc.dg/tree-ssa/ctz-int.c: Ditto.
	* gcc.dg/tree-ssa/ctz-long-long.c: Ditto.
	* gcc.dg/tree-ssa/ctz-long.c: Ditto.
	* gcc.dg/tree-ssa/ctz-ch.c: New test.
	* gcc.dg/pr41488.c: Add -fno-tree-scev-cprop.