tree-optimization/121631 - UB in vector epilogue
The vectorizer fails to take UB due to signed overflow into account
when generating code for the epilogue of a signed reduction.  The
following tries to make sure to perform the actual reduction
computations in an unsigned type.  I did not bother to adjust
inputs to internal functions like .REDUC_PLUS.

	PR tree-optimization/121631
	* tree-vect-loop.cc (vect_create_epilog_for_reduction):
	When the reduction operation invokes UB on signed overflow
	make sure to perform operations with it on an unsigned type.