Add "RejectNegative" to some options where it doesn't make sense [PR122243]
This patch adds the "RejectNegative" property to several options where
it doesn't make sense.  These are either options of the form
"name=value" rather than an on/off switch, those that are already in a
"no-" form, or options that form a mutually-exclusive set.

Also, the fhelp, ftarget-help, and fversion options that do not take
arguments ignore the "-no" prefix so that even "-fno-help" (etc)
causes help to be printed instead of suppressing help output.  Since that
behavior is not useful, I've added RejectNegative to those options as well.

gcc/analyzer/ChangeLog
	PR other/122243
	* analyzer.opt (fanalyzer-verbosity=): Add RejectNegative.

gcc/c-family/ChangeLog
	PR other/122243
	* c.opt: (fdeps-format=): Add RejectNegative.
	(fdeps-file=): Likewise.
	(fdeps-target=): Likewise.
	(Walloc-size-larger-than=): Likewise.
	(Wno-alloc-size-larger-than): Likewise.
	(Walloca-larger-than=): Likewise.
	(Wno-alloca-larger-than): Likewise.
	(Woverloaded-virtual=): Likewise.
	(Wvla-larger-than=): Likewise.
	(Wno-vla-larger-than): Likewise.
	(fopenacc-dim=): Likewise.
	(femit-struct-debug-baseonly): Likewise.
	(femit-struct-debug-reduced): Likewise.
	(femit-struct-debug-detailed=): Likewise.

gcc/ChangeLog
	PR other/122243
	* common.opt (fhelp): Add RejectNegative.
	(fhelp=): Likewise.
	(ftarget-help): Likewise.
	(fversion): Likewise.
	(Wno-frame-larger-than): Likewise.
	(Wno-larger-than): Likewise.
	(Wno-stack-usage): Likewise.
	(fdiagnostics-minimum-margin-width=): Likewise.
	(flto-incremental=): Likewise.
	(foffload=): Likewise.
	(foffload-options=): Likewise.
	(foffload-abi-host-opts=): Likewise.
	(fpatchable-function-entry=): Likewise.
	(gno-pubnames): Likewise.
	(gpubnames): Likewise.
	(ggnu-pubnames): Likewise.