ia64: convert widen_[us]sum into convert optab [PR122069]
The target does not seem to have a maintainer listed, I've CC'ed a group of
global maintainers instead hoping one of you could approve it.

This patch is a mechanical rewrite of the widen_[us]sum optabs from a direct to
a conversion optab.  The result of which requires the output mode to be added to
the existing patterns.

No change in functionality is expected.

gcc/ChangeLog:

	PR middle-end/122069
	* config/ia64/vect.md (widen_usumv8qi3): Renamed ...
	(widen_usumv4hiv8qi3): ... into this.
	(widen_usumv4hi3): Renamed ...
	(widen_usumv2siv4hi3): ... into this.
	(widen_ssumv8qi3): Renamed ...
	(widen_ssumv4hiv8qi3): ... into this.
	(widen_ssumv4hi3): Renamed ...
	(widen_ssumv2siv4hi3): ... into this.