ipa: Delete callback edges when redirecting to unreachable.
When a callback-carrying edge is redirected to __builtin_unreachable,
the associated callbacks will never get called, so the corresponding
callback edges must be deleted, as they no longer reflect the reality.

The line in analyze_function_body is an obvious typo I discovered during
debugging, so I decided to bundle it in.

gcc/ChangeLog:

	* ipa-fnsummary.cc (redirect_to_unreachable): Purge callback
	edges when redirecting the carrying edge.
	(analyze_function_body): Fix typo.

Signed-off-by: Josef Melcr <jmelcr02@gmail.com>