aarch64: Add support for menable-sysreg-checking flag.
Hi All,

In the current Binutils we have disabled the feature gating for sysreg
by default and we have introduced a new flag "-menable-sysreg-checking"
to renable some of this checking.

However in GCC, we have disabled the feature gating of sysreg to read/write
intrinsics __arm_[wr]sr* and we have not added any mechanism to check the
feature gating if needed similar to Binutils.

This patch adds the support for the flag "-menable-sysreg-checking" which
renables some of the feature checking of sysreg to read/write intrinsics
__arm_[wr]sr* similar to Binutils.

For inline assembly, sysreg checks are not performed by CC1 and are
instead delegated to the assembler. By default, the assembler does not
perform these checks either. With this patch, the -menable-sysreg-checking
flag passed to the compiler will also be propagated to the assembler,
enabling sysreg checking for inline assembly.

gcc/ChangeLog:

	* config/aarch64/aarch64-elf.h (ASM_SPEC): Update the macro.
	* config/aarch64/aarch64.cc (aarch64_valid_sysreg_name_p):
	Add feature check condition.
	(aarch64_retrieve_sysreg): Likewise.
	* config/aarch64/aarch64.opt (menable-sysreg-checking):
	Define new flag.
	* doc/invoke.texi (menable-sysreg-checking): Document new flag.

gcc/testsuite/ChangeLog:

	* gcc.target/aarch64/acle/asm-inlined-sysreg-1.c: New test.
	* gcc.target/aarch64/acle/asm-inlined-sysreg-2.c: Likewise.
	* gcc.target/aarch64/acle/rwsr-gated-1.c: Likewise.
	* gcc.target/aarch64/acle/rwsr-gated-2.c: Likewise.
	* lib/target-supports.exp
	(check_effective_target_aarch64_sysreg_guarding_ok): Check
	assembler support of -menable-sysreg-checking flag.