c++: Implement C++23 P2674R1 - A trait for implicit lifetime types
The following patch attempts to implement the compiler side of the
C++23 P2674R1 paper.  As mentioned in the paper, since CWG2605
the trait isn't really implementable purely on the library side.

Because it is implemented completely on the compiler side, it
just uses SCALAR_TYPE_P and so can e.g. accept __int128 even in
-std=c++23 mode, even when std::is_scalar_v<__int128> is false in
that case.  And as an extention it (like Clang) accepts _Complex
types and vector types.
I must say I'm quite surprised that any array types are considered
implicit-lifetime, even if their element type is not, but perhaps
there is some reason for that.
Because std::is_array_v<int[0]> is false, it returns false for that
as well, dunno if that shouldn't be changed for implicit-lifetime.
It accepts also VLAs.

The library part has been split into a separate patch still pending
review; committing it now so that reflection can use it in its
std::meta::is_implicit_lifetime_type implementation.

2025-10-21  Jakub Jelinek  <jakub@redhat.com>

gcc/cp/
	* cp-tree.h: Implement C++23 P2674R1 - A trait for implicit lifetime
	types.
	(implicit_lifetime_type_p): Declare.
	* tree.cc (implicit_lifetime_type_p): New function.
	* cp-trait.def (IS_IMPLICIT_LIFETIME): New unary trait.
	* semantics.cc (trait_expr_value): Handle CPTK_IS_IMPLICIT_LIFETIME.
	(finish_trait_expr): Likewise.
	* constraint.cc (diagnose_trait_expr): Likewise.
gcc/testsuite/
	* g++.dg/ext/is_implicit_lifetime.C: New test.