ada: Restore alignment check for address clause given with alignment clause
The alignment check for address clauses had historically been performed
universally by GNAT, until a decade ago when it was disabled by default
on non-strict-alignment platforms.

This seems questionable when an alignment clause is also given for the
entity since the intent is clear, so the change restores it as well as
its associated warning.

Moreover, this alignment check had historically been subsumed into the
language-defined Range_Check, until two decades ago when it became the
implementation-defined Alignment_Check, while still being part of the
Range_Check to ease the transition.  Now is probably a good time to cut
the apron strings and make Alignment_Check standalone.

gcc/ada/ChangeLog:

	* doc/gnat_rm/representation_clauses_and_pragmas.rst (Address
	Clauses): Adjust description of alignment checks.
	* sem_ch13.adb (Address_Clause_Check_Record): Remove now unused
	Alignment_Checks_Suppressed component.
	(Alignment_Checks_Suppressed): Delete.
	(Analyze_Attribute_Definition_Clause) <Attribute_Alignment>: Do not
	consider the status of range checks to decide whether to generate an
	alignment check.  Always generate it if an alignment clause is also
	given for the entity.
	<Attribute_Alignment>: Generate an alignment check if an address
	clause is also given for the entity.
	(Register_Address_Clause_Check): Adjust for above removal.
	(Validate_Address_Clauses): Test the Check_Address_Alignment flag
	on the clause to decide whether to give the warning.
	* gnat_rm.texi: Regenerate.
	* gnat_ugn.texi: Regenerate.