Only document -A/--assert options in cpp manual [PR122243]
Assertions are a preprocessor feature that has been declared obsolete
with strong warnings not to use them since 2001.  The main GCC manual
documents the -A command-line option but doesn't include the section
that explains the purpose of the feature or that it is obsolete; that
material appears only in the preprocessor manual.  It seems rather pointless
to clutter up the GCC manual with unhelpful documentation of an obsolete
feature, so I've restricted the option documentation to the
preprocessor manual too.  I've also added the missing documentation
entries for the long form of the option, --assert.

gcc/ChangeLog
	PR other/122243
	* doc/cppopts.texi (-A): Restrict option documentation to the CPP
	manual.  Also document the --assert form.
	* doc/invoke.texi (Option Summary): Don't list the -A option.