vect: Add support for boolean reductions for VLA
The support for the new boolean reduction optabs didn't quite work for VLA
because the code later on insists on the target still having a shift-and-insert
optab.

This is however not needed if the target can do the reduction using the new
optabs, and the initial reduction value matches the neutral value and we
have one SLP lane while not having a reduction chain.

gcc/ChangeLog:

	* tree-vect-loop.cc (vectorizable_reduction): Don't always require
	IFN_VEC_SHL_INSERT when using reduc sbool optabs.