AArch64: Add support for boolean reductions for SVE
The vectorizer has learned how to do boolean reductions of masks to a C bool
for the operations OR, XOR and AND.

This implements the new optabs for SVE.

For SVE & and the | case would use the CC registers.

or_reduc:
        ptest   p0, p0.b
        cset    w0, any

and_reduc:
        ptrue   p3.b, all
        nots    p3.b, p3/z, p0.b
        cset    w0, none

and the ^ case we'd see if the number of active predicate lanes
is a multiple of two.

xor_reduc:
        ptrue   p3.b, all
        cntp    x0, p3, p0.b
        and     w0, w0, 1

gcc/ChangeLog:

	* config/aarch64/aarch64-sve.md (reduc_sbool_and_scal_<mode>,
	reduc_sbool_ior_scal_<mode>, reduc_sbool_xor_scal_<mode>): New.

gcc/testsuite/ChangeLog:

	* gcc.target/aarch64/sve/vect-reduc-bool-1.c: New test.
	* gcc.target/aarch64/sve/vect-reduc-bool-2.c: New test.
	* gcc.target/aarch64/sve/vect-reduc-bool-3.c: New test.
	* gcc.target/aarch64/sve/vect-reduc-bool-4.c: New test.
	* gcc.target/aarch64/sve/vect-reduc-bool-5.c: New test.
	* gcc.target/aarch64/sve/vect-reduc-bool-6.c: New test.
	* gcc.target/aarch64/sve/vect-reduc-bool-7.c: New test.
	* gcc.target/aarch64/sve/vect-reduc-bool-8.c: New test.
	* gcc.target/aarch64/sve/vect-reduc-bool-9.c: New test.