xtensa: Remove redundant use of 'i'-constraint
Because it is redundant to specify 'i'-constraints on operands in single-
alternative match templates whose predicates are "const_int_operand" itself
or those that imply CONST_INT_P().

This patch also removes the 'i'-constraints on the next argument of the
callee (the number of bytes of arguments) in the four "call_internal"
patterns, since we are not interested in these arguments.

gcc/ChangeLog:

	* config/xtensa/xtensa.md (*addsubx, *subsi3_from_const,
	*xtensa_clamps, *andsi3_const_pow2_minus_one,
	*andsi3_const_negative_pow2, *andsi3_const_shifted_mask,
	*splice_bits, extvsi_internal, extzvsi_internal,
	*extzvsi-1bit_ashlsi3, *extzvsi-1bit_addsubx, insvsi, *lsiu, *ssiu,
	*lsip, *ssip, *shift_per_byte_omit_AND_0, *shift_per_byte_omit_AND_1,
	*shlrd_const, *shlrd_per_byte_omit_AND, *masktrue_const_bitcmpl,
	*masktrue_const_pow2_minus_one, *masktrue_const_negative_pow2,
	*masktrue_const_shifted_mask, call_internal, call_value_internal,
	sibcall_internal, sibcall_value_internal, entry,
	*eqne_zero_masked_bits, *eqne_in_range): Remove 'i'-constraint.