Fortran: fix TRANSFER of subarray component references [PR122386]
Commit r16-518 introduced a change that fixed inquiry references of complex
arrays as argument to the TRANSFER intrinsic by forcing a temporary.  The
solution taken however turned out not to be generalizable to component
references of nested derived-type arrays.  A better way is the revert that
patch and force the generation of a temporary when the SOURCE expression is
a not simply-contiguous array.

	PR fortran/122386

gcc/fortran/ChangeLog:

	* dependency.cc (gfc_ref_needs_temporary_p): Revert r16-518.
	* trans-intrinsic.cc (gfc_conv_intrinsic_transfer): Force temporary
	for SOURCE not being a simply-contiguous array.

gcc/testsuite/ChangeLog:

	* gfortran.dg/transfer_array_subref_2.f90: New test.

(cherry picked from commit 2febf3b968329aceeeea7805af98ed98a8c67e75)